# docs/swagger.yaml
basePath: /
definitions:
...
  dto.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
...
  dto.ErrorResponse:
    properties:
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.OAuthURLResponse:
    properties:
      provider:
        type: string
      state:
        type: string
      url:
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      address:
        type: string
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified_at:
        type: string
      full_name:
        type: string
      id:
        type: integer
      phone:
        type: string
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: API documentation for Foods & Drinks.
  title: Foods & Drinks API
  version: "1.0"
paths:
...
  /api/v1/auth/oauth/{provider}:
    get:
      description: Get OAuth authorization URL for the specified provider
      parameters:
      - description: OAuth provider (google, facebook, twitter)
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OAuthURLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Initiate OAuth flow
      tags:
      - oauth
  /api/v1/auth/oauth/{provider}/callback:
    get:
      description: Handle OAuth callback from provider
      parameters:
      - description: OAuth provider (google, facebook, twitter)
        in: path
        name: provider
        required: true
        type: string
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: State for CSRF protection
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Handle OAuth callback
      tags:
      - oauth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"